Renamed ImageController to ImageViewerController.
